requirements.txt
numpy
pandas
seaborn
matplotlib.pyplot
schickit-learn == 1.5.2
imbalanced-learn == 0.12.4
lightgbm == 4.5.0
xgboost == 2.1.2
catboost == 1.2.7
optuna == 4.0.0
dvc == 3.56.0
mlflow == 2.17.2
dotenv == 1.0.1
dagshub == 0.3.42
fastapi == 0.115.5
uvicorn == 0.32.1
pydantic == 2.9.2
multipart == 0.0.17